social/tests/requirements-python3.txt
httpretty==0.6.5
coverage>=3.6
mock==1.0.1
nose>=1.2.1
requests>=1.1.0
PyJWT==0.2.1
unittest2py3k==0.5.1
